public/Projects/Rehanuddin Shaik/AI_Image_Analyser/code/requirements.txt
Pillow
requests
streamlit
numpy
opencv-python-headless
Pillow
pandas
